fix(execution): able to create multiple heredocs, however there is a issue with dup2 with multiple pipes
