+++ src/parsing/parse.c
@@ -6,7 +6,7 @@
 /*   By: ryannnaa <ryannnaa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/25 17:20:28 by tiatan            #+#    #+#             */
-/*   Updated: 2025/03/05 22:07:32 by ryannnaa         ###   ########.fr       */
+/*   Updated: 2025/03/06 17:41:28 by ryannnaa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,10 @@ void	parse_redir(t_node **token, t_redirect **ast_rd)
 	new_rd->type = (*token)->type;
 	(*token) = (*token)->next;
 	if (new_rd->type == T_HEREDOC)
+	{
 		new_rd->file = rd_expand((*token)->value, 0);
+		new_rd->fd = create_heredoc(new_rd->file);
+	}
 	else
 		new_rd->file = rd_expand((*token)->value, 1);
 	if (!(*ast_rd))
